# api-docs.yaml
openapi: 3.0.0
info:
  title: API DOCS PPOB
  version: 1.0.0
servers:
  - url: http://34.101.78.228:2424/api/v1
  - url: https://github.com/darulfh/skuy_pay_be.kucinghitam.tech/api/v1
paths:
...
  /user/ppd:
    get:
      tags:
        - Pulsa Paket Data (PPD)
      summary: Get PPD by user
      description: Get pulsa paket data by user
      operationId: getUserPPD
      parameters:
        - in: query
          name: type
          required: true
          schema:
            type: string
            enum:
              - pulsa
              - paket_data
            example: pulsa
        - in: query
          name: provider
          required: true
          schema:
            type: string
            example: Telkomsel
        - in: query
          name: limit
          schema:
            type: integer
            example: 10
        - in: query
          name: page
          schema:
            type: integer
            example: 1
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successfully get all ppd
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AllPPDResponse"

        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Metadata"
              example:
                status: 401
                message: token unauthorized

        404:
          description: ppd not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Metadata"
              example:
                status: 404
                message: ppd not found
    post:
      tags:
        - Pulsa Paket Data (PPD)
      summary: Create Transaction PPD
      description: Create Transaction PPD
      operationId: createTransactionPPD
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransactionPPDRequest"
      responses:
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Metadata"
              example:
                status: 401
                message: token unauthorized

        404:
          description: ppd not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Metadata"
              example:
                status: 404
                message: ppd not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  # parameters:
  schemas:
...
    AllPPDResponse:
      type: object
      properties:
        metadata:
          $ref: "#/components/schemas/Metadata"
        data:
          type: array
          items:
            $ref: "#/components/schemas/PPD"
...
    TransactionPPDRequest:
      type: object
      properties:
        type:
          type: string
          enum:
            - paket_data
            - pulsa
          example: "paket_data"
        product_id:
          type: string
          example: "e56d5697-e412-4fd7-968c-d4cc9a6d7317"
        discount_id:
          type: string
          example: "87c3b410-6aa5-4028-83e0-a30f2e36270e"
        phone_number:
          type: string
          example: "08132132132"
...
    Metadata:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
...
    PPD:
      type: object
      properties:
        name:
          type: string
          example: Internet Bulanan OMG 100rb
        type:
          type: string
          default: pulsa
          enum: [pulsa, paket_data]

        code:
          type: string
          example: TSOMG75
        provider:
          type: string
          example: Telkomsel
        price:
          type: number
          example: 100000
        is_active:
          type: boolean
          example: true
        description:
          type: string
          example: Kouta Utama 20 GB, Kouta malam 5GB
...